image-processing-module/process-images/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
astor==0.8.1
async-timeout==4.0.2
attrdict==2.0.1
attrs==22.1.0
Babel==2.11.0
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
colorlog==6.7.0
commonmark==0.9.1
contourpy==1.0.6
cssselect==1.2.0
cssutils==2.6.0
cycler==0.11.0
Cython==0.29.32
datasets==2.8.0
decorator==5.1.1
dill==0.3.4
easyocr==1.6.2
et-xmlfile==1.1.0
fastapi==0.88.0
filelock==3.8.2
fire==0.5.0
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
future==0.18.2
gast==0.4.0
h11==0.14.0
huggingface-hub==0.11.1
idna==3.4
imageio==2.23.0
imgaug==0.4.0
importlib-metadata==5.2.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
Levenshtein==0.20.8
lmdb==1.4.0
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
multidict==6.0.3
multiprocess==0.70.12.2
networkx==2.8.8
ninja==1.11.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-contrib-python==4.4.0.46
opencv-python==4.6.0.66
opencv-python-headless==4.5.4.60
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddle2onnx==1.0.5
paddlefsl==1.1.0
paddlenlp==2.4.5
paddleocr==2.4
paddlepaddle==2.1.3
pandas==1.5.2
pdf2docx==0.5.6
Pillow==9.2.0
premailer==3.10.0
protobuf==3.20.0
pyarrow==10.0.1
pyclipper==1.3.0.post4
pycryptodome==3.16.0
pydantic==1.10.2
Pygments==2.13.0
PyMuPDF==1.20.2
pyparsing==3.0.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-Levenshtein==0.20.8
pytz==2022.7
PyWavelets==1.4.1
PyYAML==6.0
rapidfuzz==2.13.7
redis==4.4.0
requests==2.28.1
responses==0.18.0
rich==12.6.0
scikit-image==0.19.3
scikit-learn==1.2.0
scipy==1.9.3
sentencepiece==0.1.97
seqeval==1.2.2
shapely==2.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
termcolor==2.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
torch==1.13.1
torchvision==0.14.1
tqdm==4.64.1
typer==0.7.0
typing-extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
visualdl==2.4.1
Werkzeug==2.2.2
xxhash==3.1.0
yarl==1.8.2
zipp==3.11.0
